@@ -1,9 +1,11 @@
 module github.com/uhppoted/uhppoted-rest
 
-go 1.21
+go 1.22
+
+toolchain go1.22.0
 
 require (
-	github.com/uhppoted/uhppote-core v0.8.7
-	github.com/uhppoted/uhppoted-lib v0.8.7
-	golang.org/x/sys v0.15.0
+	github.com/uhppoted/uhppote-core v0.8.8-0.20240215183313-d631a62bd158
+	github.com/uhppoted/uhppoted-lib v0.8.8-0.20240216164353-6533d27adfed
+	golang.org/x/sys v0.17.0
 )
